dreambox-boottool-cf: Add root_to_cf script, add to 7025 feed
The script must be run manually, to prepare the CF card. This means
that if you don't, you may end up killing the box and need to reflash...

+++ recipes/dreambox/dreambox-boottool-cf.bb
@@ -4,10 +4,10 @@ PRIORITY = "optional"
 MAINTAINER = "Mike Looijmans <MiLo@pli-images.org>"
 
 PV = "1.0"
-PR = "r1.3"
+PR = "r2"
 DEPENDS = "klcc-cross"
 
-SRC_URI = "file://boottool-${MACHINE}.c"
+SRC_URI = "file://boottool-${MACHINE}.c file://root_to_cf.sh"
 
 S = "${WORKDIR}/"
 
@@ -15,6 +15,8 @@ do_install_append() {
 	install -d ${D}/boot/bin
 	install ${S}/boottool ${D}/boot/bin/initcf
 	install -d ${D}/boot/dev
+	install -d ${D}/usr/bin
+	install -m 755 ${WORKDIR}/root_to_cf.sh ${D}/usr/bin/
 }
 
 do_compile_append() {
@@ -24,9 +26,20 @@ do_compile_append() {
 pkg_postinst() {
 	mknod -m 660 $D/boot/dev/hdc b 22 0
 	mknod -m 660 $D/boot/dev/hdc1 b 22 1
-	mv $D/boot/bin/init $D/boot/bin/initflash
-	ln $D/boot/bin/initcf $D/boot/bin/init
+	if [ ! -f $D/boot/bin/initflash ]
+	then
+		mv $D/boot/bin/init $D/boot/bin/initflash
+	fi
+	ln -f $D/boot/bin/initcf $D/boot/bin/init
+}
+
+pkg_prerm() {
+	if [ -f /boot/bin/initflash ]
+	then
+		rm -f /boot/bin/init
+		mv /boot/bin/initflash /boot/bin/init
+	fi
 }
 
 PACKAGE_ARCH := "${MACHINE_ARCH}"
-FILES_${PN} = "/boot/bin/initcf /boot/dev"
+FILES_${PN} = "/boot/bin/initcf /boot/dev /usr/bin"
